services/receipt-scanner/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10
pydantic==2.5.0
python-jose[cryptography]==3.3.0